created new app.py with multiple model handling capability

+++ app.py
@@ -7,9 +7,18 @@ from tensorflow.keras.preprocessing import image
 
 app = Flask(__name__)
 
-model_vgg16 = load_model('models/detecting_freshness_of_fish_cube_VGG16.h5')
-classes = ['Healthy', 'Unhealthy']
+try:
+    fish_type_model = load_model('models/detecting_freshness_of_fish_cube_VGG16.h5')
+    healty_unhealthy_yellowfin_model = load_model('models/detecting_freshness_of_fish_cube_VGG16.h5')
+    healty_unhealthy_skipjack_model = load_model('models/detecting_freshness_of_fish_cube_VGG16.h5')
+except Exception as e:
+    print(e)
+    exit(0)
 
+fish_type_classes = ['Yellofin_tuna', 'Skipjack_tuna', "Invalid_image"]
+healty_unhealthy_classes = ['Healthy', 'Unhealthy']
+
+img_path = 'temp/temp_image.jpg'
 
 def add_headers(output):
     response = make_response(jsonify(output))
@@ -31,6 +40,31 @@ def prepare(img_path):
     return np.expand_dims(x, axis=0)
 
 
+def detect_fish_type(img_url):
+    
+    download_image(img_url, img_path)
+    img_data = prepare(img_path)
+    result_vgg16 = fish_type_model.predict(img_data)
+    class_result = np.argmax(result_vgg16, axis=1)
+    prediction = fish_type_classes[class_result[0]]
+
+    return [prediction, imge_data]
+
+
+def detect_healty_unhealthy(fish_type, img_data):
+
+    model = healty_unhealthy_skipjack_model
+    if(fish_type==fish_type_classes[0]):
+        #select the yellofin model
+        model = healty_unhealthy_yellowfin_model
+
+    result_vgg16 = model.predict(img_data)
+    class_result = np.argmax(result_vgg16, axis=1)
+    prediction = healty_unhealthy_classes[class_result[0]]
+
+    return prediction
+
+
 @app.route('/')
 def index():
     return render_template("index.html")
@@ -39,18 +73,19 @@ def index():
 @app.route('/predict', methods=['POST'])
 def predict():
     try:
-        img_url = request.json['img_url']
-        img_path = 'temp/temp_image.jpg'
-        download_image(img_url, img_path)
-        img_data = prepare(img_path)
-        result_vgg16 = model_vgg16.predict(img_data)
-        class_result = np.argmax(result_vgg16, axis=1)
-        prediction = classes[class_result[0]]
-        os.remove(img_path)
-        return add_headers({'prediction': prediction})
+        fish_type_data = detect_fish_type(request.json['img_url'])
+
+        if(fish_type==fish_type_classes[2]):
+            return add_headers({'error': fish_type})
+        else:
+            try:
+                prediction = detect_healty_unhealthy(fish_type_data[0], fish_type_data[1])
+                os.remove(img_path)
+                return add_headers({'prediction': prediction})
+            except Exception as e:
+                return add_headers({'error': str(e)})
     except Exception as e:
         return add_headers({'error': str(e)})
-
 
 if __name__ == '__main__':
     from waitress import serve
